Portability fix: Change unsigned long to size_t in some hash functions. When buildifn for Windows, add windows_compatibilty.h to hash build.

+++ src/googleapis/util/hash.cc
@@ -61,11 +61,11 @@ uint32 Hash32StringWithSeedReferenceImplementation(const char *s, size_t len,
   return n;
 }
 
-uint32 Hash32StringWithSeed(const char *s, uint32 len, uint32 c) {
+uint32 Hash32StringWithSeed(const char *s, size_t len, uint32 c) {
   return Hash32StringWithSeedReferenceImplementation(s, len, c);
 }
 
-uint64 Hash64StringWithSeed(const char *s, unsigned long len, uint64 seed) {
+uint64 Hash64StringWithSeed(const char *s, size_t len, uint64 seed) {
   uint64 n = seed;
   size_t prime1 = 0, prime2 = 8;  // Indices into kPrimes64
   union {
